feat: Expose available actions for typedefs
We don't yet expose actions for constructor fields (i.e. `Available.forTypeDefConsFieldNode` always returns `[]`), since making this work will, unlike the other positions, require changes to the core of the library.

We _could_ actually make all the `for*`s part of one definition, now that `Selection` is in `App.Base`. But we do actually use the individual functions in some tests, and with them separate, we have slightly more control, in that we don't need to provide as much context.

Note that most of the changes in this commit are actually knock-on effects of generalising `Selection` to cover type defs.

Signed-off-by: George Thomas <georgefsthomas@gmail.com>